refactor all functions regarding bidding functionality. Moved some other functions/files inside repo. add jsdocs to some functions
