+++ src/js/shared/handlers/auctions/auctionDetailsPageHandler.mjs
@@ -8,7 +8,7 @@ export const auctionDetailsPageHandler = async () => {
   const auctionId = getQueryParameter("id");
 
   if (!auctionId) {
-    // window.location.href = "../auctions/";
+    window.location.href = "../auctions/";
   }
 
   const endpoint = `auction/listings/${auctionId}`;
@@ -24,7 +24,7 @@ export const auctionDetailsPageHandler = async () => {
     const singleAuctionData = data?.data;
 
     if (!singleAuctionData) {
-      // window.location.href = "../auctions";
+      window.location.href = "../auctions";
     }
 
     renderCompleteDetailsPage(singleAuctionData, container);
