+++ src/js/shared/handlers/auctions/handleNewListing.mjs
@@ -1,6 +1,6 @@
-import { createListing } from "../../api/new-listing/createListing.mjs";
-import { validateNewListingForm } from "../../ui/auctions/new-listing/validateNewListingForm.mjs";
-import { getNewListingFormData } from "../../utils/new-listing/getNewListingFormData.mjs";
+import { createListing } from "../../api/auctions/createListing.mjs";
+import { validateNewListingForm } from "../../ui/auctions/validateNewListingForm.mjs";
+import { getNewListingFormData } from "../../utils/auctions/getNewListingFormData.mjs";
 
 export const handleNewListing = async () => {
   const form = document.querySelector("#create-listing-form");
