+++ src/js/shared/listeners/auctions/submitBid.mjs
@@ -1,0 +1,10 @@
+import { handleBiddingForm } from "../../handlers/auctions/handleBiddingForm.mjs";
+
+export const submitBid = () => {
+  document.addEventListener("submit", (e) => {
+    if (e.target.id === "bid-form") {
+      e.preventDefault();
+      handleBiddingForm("#message", e);
+    }
+  });
+};
