+++ src/js/shared/utils/auctions/getNewListingFormData.mjs
